group1_baseline/requirements-core.txt
numpy
Pillow
transformers
huggingface_hub
datasets
jax
flax
optax
humanize
